♻️ Refactor filekit-compose to match this new API
