+++ filekit-compose/src/nonWebMain/kotlin/io/github/vinceglb/filekit/compose/FileKitCompose.nonWeb.kt
@@ -1,0 +1,50 @@
+package io.github.vinceglb.filekit.compose
+
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.getValue
+import androidx.compose.runtime.remember
+import androidx.compose.runtime.rememberCoroutineScope
+import androidx.compose.runtime.rememberUpdatedState
+import io.github.vinceglb.filekit.core.FileKit
+import io.github.vinceglb.filekit.core.FileKitPlatformSettings
+import io.github.vinceglb.filekit.core.PlatformFile
+import io.github.vinceglb.filekit.core.pickDirectory
+import kotlinx.coroutines.launch
+
+@Composable
+public fun rememberDirectoryPickerLauncher(
+    title: String? = null,
+    initialDirectory: String? = null,
+    platformSettings: FileKitPlatformSettings? = null,
+    onResult: (PlatformFile?) -> Unit,
+): PickerResultLauncher {
+    // Init FileKit
+    InitFileKit()
+
+    // Coroutine
+    val coroutineScope = rememberCoroutineScope()
+
+    // Updated state
+    val currentTitle by rememberUpdatedState(title)
+    val currentInitialDirectory by rememberUpdatedState(initialDirectory)
+    val currentOnResult by rememberUpdatedState(onResult)
+
+    // FileKit
+    val fileKit = remember { FileKit }
+
+    // FileKit launcher
+    val returnedLauncher = remember {
+        PickerResultLauncher {
+            coroutineScope.launch {
+                val result = fileKit.pickDirectory(
+                    title = currentTitle,
+                    initialDirectory = currentInitialDirectory,
+                    platformSettings = platformSettings,
+                )
+                currentOnResult(result)
+            }
+        }
+    }
+
+    return returnedLauncher
+}
